+++ tests/testthat/test-templates.R
@@ -51,10 +51,9 @@ verify_models <- function(model, prefix, tune, verbose) {
 test_that("all model templates", {
   local_edition(3)
 
-  models <- c(
-    "glmnet", "xgboost", "ranger", "kknn", "earth", "cubist",
-    "kernlab_svm_poly", "kernlab_svm_rbf", "C5.0"
-  )
+  models <- c("bag_tree_rpart", "C5.0", "cubist", "dbarts", "earth", "glmnet",
+              "kernlab_svm_poly", "kernlab_svm_rbf", "kknn", "mgcv", "mixOmics",
+              "nnet", "ranger", "rpart", "xgboost", "xrf")
 
   test_config <-
     expand.grid(