+++ NEWS.md
@@ -1,6 +1,9 @@
 # usemodels (development version)
 
+* Added `use_nnet()`, `use_rpart()`, `use_bag_tree_rpart()`, `use_mgcv()`, `use_dbarts()`, `use_mixOmics()`, `use_xrf()`. 
+
 * Fix `recipe()` call when `clipboard = TRUE`
+
 
 # usemodels 0.2.0